requirements.txt
flask==3.1.0
requests==2.32.3
python-dotenv==1.1.0
gunicorn==23.0.0
openai>=1.30.0
notion-client>=2.2.0
google-api-python-client>=2.130.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
tweepy>=4.14.0
